+++ src/tui/tcell.go
@@ -203,6 +203,12 @@ func (r *FullscreenRenderer) Refresh() {
 	// noop
 }
 
+// TODO: Pixel width and height not implemented
+func (r *FullscreenRenderer) Size() TermSize {
+	cols, lines := _screen.Size()
+	return TermSize{lines, cols, 0, 0}
+}
+
 func (r *FullscreenRenderer) GetChar() Event {
 	ev := _screen.PollEvent()
 	switch ev := ev.(type) {
@@ -541,6 +547,7 @@ func fill(x, y, w, h int, n ColorPair, r rune) {
 }
 
 func (w *TcellWindow) Erase() {
+	w.drawBorder(false)
 	fill(w.left-1, w.top, w.width+1, w.height-1, w.normal, ' ')
 }
 
@@ -690,6 +697,10 @@ func (w *TcellWindow) CFill(fg Color, bg Color, a Attr, str string) FillReturn {
 		bg = w.normal.Bg()
 	}
 	return w.fillString(str, NewColorPair(fg, bg, a))
+}
+
+func (w *TcellWindow) DrawBorder() {
+	w.drawBorder(false)
 }
 
 func (w *TcellWindow) DrawHBorder() {