+++ CHANGELOG.md
@@ -1,9 +1,23 @@
 CHANGELOG
 =========
 
+0.43.1
+------
+- (Experimental) Sixel image support in preview window (not available on Windows)
+    - `$FZF_PREVIEW_PIXEL_WIDTH` and `$FZF_PREVIEW_PIXEL_HEIGHT` are set to
+      the pixel width and height of the preview window
+    - [bin/fzf-preview.sh](bin/fzf-preview.sh) is added to demonstrate how to
+      display an image using Kitty image protocol or Sixel. You can use it
+      like so:
+      ```sh
+      fzf --preview='fzf-preview.sh {}'
+      ```
+- Bug fixes
+
 0.43.0
 ------
 - (Experimental) Added support for Kitty image protocol in the preview window
+  (not available on Windows)
   ```sh
   fzf --preview='
     if file --mime-type {} | grep -qF image/; then